+++ examples/esbuild-with-plugins/lambda.ts
@@ -1,0 +1,4 @@
+import capitalize from "https://cdn.skypack.dev/lodash.capitalize";
+
+export const handler = async (input: { body: string }) =>
+  capitalize(input?.body);